Ensure all stops have valid latitude and longitude
As the Easting and Northing values are mandatory (but Latitude and Longitude are note), I came across stops which couldn't be mapped with folium (needs lat long). As such, I inserted an extra step to ensure that all stops have Latitude and Longitude calculate by converting their Easting and Northing.